fixes in generate features for action_tags

+++ NNModels/generate/generate_database_fight.py
@@ -268,7 +268,7 @@ range_stops = np.array(range_stops).astype(np.int32)
 contact_states = np.concatenate(contact_states, axis=0).astype(np.uint8)
 
 if ms.animation_type == "actions":
-    action_tags = np.concatenate(action_tags, axis=0).astype(np.uint8)
+    action_tags = np.concatenate(action_tags, axis=0).astype(np.int32)
 
 """ Write Database """
